+++ VisioAutomation_2010/VisioAutomation/VisioAutomation.csproj
@@ -127,6 +127,7 @@
     <Compile Include="ShapeSheet\CellGroups\SrcValuePair.cs" />
     <Compile Include="ShapeSheet\ShapeSheetHelper.cs" />
     <Compile Include="Shapes\UserDefinedCellDictionary.cs" />
+    <Compile Include="Shapes\UserDefinedCellKeyValuePair.cs" />
     <Compile Include="SurfaceTarget.cs" />
     <Compile Include="Shapes\CustomPropertyDictionary.cs" />
     <Compile Include="Shapes\HyperlinkCells.cs" />
